Mobt 914 env update nowcasting (#2142)
* Ensures precision of output data are float32

* Reduces precision requirement of unit tests from 7 decimal places to 6.

* Fixes unit tests that were passing the wrong arguments to the methods they were testing

+++ improver/nowcasting/lightning.py
@@ -443,7 +443,7 @@ class NowcastLightning(PostProcessingPlugin):
                             data_range=(0.0, 1.0),
                             scale_range=ice_scaling,
                             clip=True,
-                        ),
+                        ).astype(np.float32),
                         cube_slice.data,
                     )
             new_cube_list.append(cube_slice)
